Fixed: Update request headers on making the request
This is really just a prelude to making the request connect and then forging a response object from the response.
